improvements based on initial PR feedback
