+++ src/Route.js
@@ -1,0 +1,3 @@
+export function Route(props) {
+  return props
+}
