+++ src/index.js
@@ -1,2 +1,23 @@
-export { router, Router, Route } from "./router"
 export { Link } from "./Link"
+export { Route } from "./Route"
+export { Router } from "./Router"
+
+// router module
+export const router = {
+  actions: {
+    set: function (state, actions, data) {
+      return data
+    },
+    go: function (state, actions, path) {
+      if (location.pathname + location.search !== path) {
+        history.pushState({}, "", path)
+        actions.set({ path: path })
+      }
+    }
+  },
+  init(state, actions) {
+    addEventListener("popstate", function () {
+      actions.set({})
+    })
+  }
+}
